Timestep passed to Update Functions and Code cleanup
~ Passing Timestep throu the update functions instead using TIme::GetTimestep();
~ Cleaned up the code as per suggestion found by ReSharper inspection and more.
+ Added Color constructor that takes double.

+++ Hazel/src/Hazel/Core/Random.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "Base.h"
-
 #include <glm/glm.hpp>
 #include <random>
 
@@ -17,7 +15,7 @@ namespace Hazel
 		/// Initialize the Random machine to a specific seed
 		/// </summary>
 		/// <param name="seed">seed</param>
-		static void Init(int seed = std::time(nullptr));
+		static void Init(int seed = static_cast<int>(std::time(nullptr)));
 
 		/// <summary>
 		/// Get random float between 0.0f - 1.0f
